+++ Core/Inc/task_manager.h
@@ -15,7 +15,7 @@ void init_tasks(
 
 void manage_tasks();
 
-void task_handle_rotary_change(uint16_t value);
+void task_handle_rotary_change(int32_t value);
 
 void task_handle_interrupt(uint8_t interrupt_event);
